# api/openapi.yaml
openapi: 3.1.0
info:
  title: Xminds Release Platform API
  version: 0.1.0-p0
  description: |
    Xminds Release Platform 管理面与工作负载 API 基础契约。
    管理人员使用当前启用 IAM OIDC 来源签发的 Bearer Token，自动化工作负载使用独立发行者与令牌。
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: 当前部署来源
security:
  - oidcBearer: []
...
paths:
...
  /api/v1/products:
    post:
      operationId: registerProduct
      summary: 注册产品及默认通道
      description: 仅平台管理员可调用；产品、默认通道和审计事件在同一数据库事务中提交。
      tags: [Products]
      x-required-action: product.register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductManifestV1'
      responses:
        '201':
          description: 产品已注册
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 产品 ID 或 Manifest 摘要已存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: Manifest 超过大小限制
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '415':
          description: 请求媒体类型不受支持
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Manifest 不符合 v1 契约
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      operationId: listProducts
      summary: 查询当前身份有权访问的产品
      tags: [Products]
      x-required-action: product.read
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: 服务端返回的不透明分页游标。
          schema:
            type: string
            maxLength: 512
      responses:
        '200':
          description: 产品页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          description: 分页参数无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/products/{product_id}/artifact-uploads:
    post:
      operationId: beginArtifactUpload
      summary: 创建可恢复制品上传会话
      description: 创建有效期 24 小时的临时上传会话；最终对象仅在服务端 SHA-256 校验通过后可见。
      tags: [Artifacts]
      security:
        - oidcBearer: []
        - workloadBearer: []
      x-required-action: artifact.publish
      parameters:
        - $ref: '#/components/parameters/ProductIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeginArtifactUploadRequest'
      responses:
        '201':
          description: 上传会话已创建
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactUpload'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 产品不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: 请求或制品声明超过大小限制
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '415':
          description: 请求媒体类型不受支持
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: 制品元数据无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/products/{product_id}/artifact-uploads/{upload_id}/parts/{part_number}:
    put:
      operationId: putArtifactUploadPart
      summary: 上传或重传制品分块
      description: 相同 part_number 可安全重传并替换旧分块；除最终分块外每块至少 5 MiB，单块不超过 256 MiB。
      tags: [Artifacts]
      security:
        - oidcBearer: []
        - workloadBearer: []
      x-required-action: artifact.publish
      parameters:
        - $ref: '#/components/parameters/ProductIDPath'
        - $ref: '#/components/parameters/UploadIDPath'
        - name: part_number
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10000
        - name: X-Part-SHA256
          in: header
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{64}$'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: 分块已保存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactUploadPart'
        '400':
          description: 上传 ID 或分块编号无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 上传会话不存在或不属于路径产品
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '410':
          description: 上传会话已过期
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '411':
          description: 缺少 Content-Length
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: 分块超过 256 MiB
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '415':
          description: 请求媒体类型不受支持
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: 分块大小或摘要无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/products/{product_id}/artifact-uploads/{upload_id}/complete:
    post:
      operationId: completeArtifactUpload
      summary: 完成上传并校验制品
      description: 服务端流式计算完整对象 SHA-256；不信任客户端 ETag，摘要不一致时隔离临时对象并创建清理任务。
      tags: [Artifacts]
      security:
        - oidcBearer: []
        - workloadBearer: []
      x-required-action: artifact.publish
      parameters:
        - $ref: '#/components/parameters/ProductIDPath'
        - $ref: '#/components/parameters/UploadIDPath'
      responses:
        '200':
          description: 制品已校验并固化，或返回已有内容寻址制品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        '400':
          description: 上传 ID 无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 上传会话不存在或不属于路径产品
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 分块不完整或上传状态冲突
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '410':
          description: 上传会话已过期
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: 完整对象摘要或大小不匹配
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/products/{product_id}/artifacts/{artifact_id}:
    get:
      operationId: getArtifact
      summary: 获取产品范围内的不可变制品元数据
      tags: [Artifacts]
      security:
        - oidcBearer: []
        - workloadBearer: []
      x-required-action: product.read
      parameters:
        - $ref: '#/components/parameters/ProductIDPath'
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 制品元数据；不暴露内部对象存储键或凭据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        '400':
          description: 制品 ID 无效
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 制品不存在或未绑定到路径产品
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/products/{product}/artifacts/{sha256}:
    get:
      operationId: downloadProductArtifact
      summary: 按 SHA-256 下载产品制品
      description: 摘要必须归属于路径中的产品；支持单个 HTTP 字节范围，不接受多段范围请求。
      tags: [Public Distribution]
      security: []
      parameters:
        - $ref: '#/components/parameters/PublicProductPath'
        - name: sha256
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{64}$'
        - name: Range
          in: header
          required: false
          schema:
            type: string
            pattern: '^bytes=[0-9]*-[0-9]*$'
      responses:
        '200':
          description: 完整制品
          headers:
            Accept-Ranges:
              schema:
                type: string
                const: bytes
            Content-Disposition:
              schema:
                type: string
                const: attachment
            ETag:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: 单段范围内容
          headers:
            Content-Range:
              schema:
                type: string
            Accept-Ranges:
              schema:
                type: string
                const: bytes
            Content-Disposition:
              schema:
                type: string
                const: attachment
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '304':
          description: ETag 未变化
        '404':
          description: 产品范围内不存在该摘要
        '416':
          description: Range 无效或请求了多段范围
        '503':
          description: 制品对象暂时不可用
...
components:
  parameters:
...
    ProductIDPath:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{1,62}$'
...
    UploadIDPath:
      name: upload_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PublicProductPath:
      name: product
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{1,62}$'
...
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: OIDC-JWT
      description: 面向控制台用户的企业 OIDC 身份令牌；Issuer 来自当前 iam_login_state 精确指向的 enabled OIDC 来源，token_use 必须精确为 human。
    workloadBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 面向 CI/CD、客户端与同步执行器的短期工作负载令牌；使用独立静态受信发行者，token_use 必须精确为 workload，不与人员验签互相回退。
  responses:
...
    Unauthorized:
      description: Bearer Token 缺失、无效或已过期
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: 身份已验证但缺少产品范围或 audit.read 权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ChannelManifest:
      type: object
      additionalProperties: false
      required: [name, display_name]
      properties:
        name:
          type: string
          pattern: '^[a-z][a-z0-9._-]{0,63}$'
        display_name:
          type: string
          minLength: 1
          maxLength: 128
    ProductManifestV1:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - product_id
        - display_name
        - artifact_types
        - version_scheme
        - compatibility_keys
        - catalog_format
        - default_channels
      properties:
        schema_version:
          type: string
          const: xminds-product-manifest/v1
        product_id:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        artifact_types:
          type: array
          minItems: 1
          maxItems: 32
          uniqueItems: true
          items:
            type: string
            pattern: '^[a-z][a-z0-9._-]{0,63}$'
        version_scheme:
          type: string
          const: semver
        compatibility_keys:
          type: array
          maxItems: 32
          uniqueItems: true
          items:
            type: string
            pattern: '^[a-z][a-z0-9._-]{0,63}$'
        catalog_format:
          type: string
          const: xminds-tuf-v1
        default_channels:
          type: array
          minItems: 1
          maxItems: 16
          items:
            $ref: '#/components/schemas/ChannelManifest'
    ProductChannel:
      type: object
      additionalProperties: false
      required: [product_id, name, display_name, position, created_at]
      properties:
        product_id:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        name:
          type: string
          pattern: '^[a-z][a-z0-9._-]{0,63}$'
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        position:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
    Product:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - schema_version
        - artifact_types
        - version_scheme
        - compatibility_keys
        - catalog_format
        - manifest
        - manifest_digest
        - status
        - channels
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        display_name:
          type: string
        schema_version:
          type: string
        artifact_types:
          type: array
          items:
            type: string
        version_scheme:
          type: string
        compatibility_keys:
          type: array
          items:
            type: string
        catalog_format:
          type: string
        manifest:
          $ref: '#/components/schemas/ProductManifestV1'
        manifest_digest:
          type: string
          pattern: '^[0-9a-f]{64}$'
        status:
          type: string
          enum: [active, inactive]
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ProductChannel'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deactivated_at:
          type: string
          format: date-time
    ProductPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        next_cursor:
          type: string
    BeginArtifactUploadRequest:
      type: object
      additionalProperties: false
      required: [artifact_type, filename, content_type, size, sha256]
      properties:
        artifact_type:
          type: string
          pattern: '^[a-z][a-z0-9._-]{0,63}$'
        filename:
          type: string
          minLength: 1
          maxLength: 255
        content_type:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          format: int64
          minimum: 1
          maximum: 21474836480
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
    ArtifactUpload:
      type: object
      additionalProperties: false
      required:
        - id
        - product_id
        - artifact_type
        - filename
        - content_type
        - expected_size
        - expected_sha256
        - status
        - expires_at
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        artifact_type:
          type: string
        filename:
          type: string
        content_type:
          type: string
        expected_size:
          type: integer
          format: int64
        expected_sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        status:
          type: string
          enum: [uploading, completed, quarantined, expired]
        artifact_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ArtifactUploadPart:
      type: object
      additionalProperties: false
      required: [upload_id, part_number, size, sha256, etag, created_at, updated_at]
      properties:
        upload_id:
          type: string
          format: uuid
        part_number:
          type: integer
          minimum: 1
          maximum: 10000
        size:
          type: integer
          format: int64
          minimum: 1
          maximum: 268435456
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        etag:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Artifact:
      type: object
      additionalProperties: false
      required:
        - id
        - product_id
        - artifact_type
        - filename
        - content_type
        - size
        - sha256
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        artifact_type:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
          minimum: 1
          maximum: 21474836480
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,127}$'
        request_id:
          type: string
          format: uuid